create httprequest class to actually parse the request and clientrequest only the info
