+++ Makefile
@@ -3,10 +3,12 @@ FLAGS = -Wall -Wextra #-Werror
 C_98 = -std=c++98
 OBJ_DIR = obj
 CPP_FILES = ServerSocket.cpp main.cpp ClientRequest.cpp ServerResponse.cpp InfoServer.cpp \
-			ServerStatusCode.cpp StringManipulations.cpp PrintingFunctions.cpp
+			ServerStatusCode.cpp StringManipulations.cpp PrintingFunctions.cpp \
+			HttpRequest.cpp
 CPP_OBJ = $(OBJ_DIR)/ServerSocket.o $(OBJ_DIR)/main.o $(OBJ_DIR)/ClientRequest.o \
           $(OBJ_DIR)/ServerResponse.o $(OBJ_DIR)/InfoServer.o $(OBJ_DIR)/ServerStatusCode.o \
-          $(OBJ_DIR)/StringManipulations.o $(OBJ_DIR)/PrintingFunctions.o
+          $(OBJ_DIR)/StringManipulations.o $(OBJ_DIR)/PrintingFunctions.o \
+		  $(OBJ_DIR)/HttpRequest.o
 NAME = server
 
 all: $(NAME)
